Allow to detect broken grids.
In some cases, the grid may be wrong and impossible to solve (for
example with two '1' in the same line in a Sudoku grid). This
commit changes the type returned by 'solve' / 'solveWithIt' in
order to be able to handle this case.

The detection of a broken grid, however, has to be implemented on
per puzzle basis. It is *not* done in this commit.