Check the grids for errors between iterations.
Consider the following 2x2 Futoshiki grid:

  1<_
  1 _

It is obviously wrong and cannot be solved. However, depending on
the order in which the solver iterates over the views, it may
select the two horizontal views first, "solves" the grid and never
see that the vertical view is wrong. This commit fixes this bug.